refactored the class hierarchy
