+++ piJEE/piJEE-ejb/src/main/java/tn/esprit/overpowered/byusforus/services/users/CompanyProfileFacade.java
@@ -14,6 +14,7 @@ import javax.persistence.PersistenceContext;
 import tn.esprit.overpowered.byusforus.entities.users.Candidate;
 import tn.esprit.overpowered.byusforus.entities.users.CompanyAdmin;
 import tn.esprit.overpowered.byusforus.entities.users.CompanyProfile;
+import tn.esprit.overpowered.byusforus.entities.users.Professional;
 
 /**
  *
@@ -33,13 +34,13 @@ public class CompanyProfileFacade extends AbstractFacade<CompanyProfile> impleme
     public CompanyProfileFacade() {
         super(CompanyProfile.class);
     }
-    List<Candidate> getSubscribersList(Long idAdmin){
+    List<Professional> getSubscribersList(Long idAdmin){
     CompanyAdmin compAdmin = em.find(CompanyAdmin.class, idAdmin);
     Long idComp = compAdmin.getCompanyProfile().getId();
     CompanyProfile compProf = em.find(CompanyProfile.class, idComp);
-        List<Candidate> listCdt = compProf.getSubscribers();
-        List<Candidate> subscribersList = new ArrayList<>();
-        for (Candidate cdtt : listCdt) {
+        List<Professional> listCdt = compProf.getSubscribers();
+        List<Professional> subscribersList = new ArrayList<>();
+        for (Professional cdtt : listCdt) {
             subscribersList.add(em.find(Candidate.class, cdtt.getId()));
         }
         return subscribersList;
