+++ piJEE/piJEE-ejb/src/main/java/tn/esprit/overpowered/byusforus/services/users/CompanyProfileFacadeRemote.java
@@ -7,8 +7,8 @@ package tn.esprit.overpowered.byusforus.services.users;
 
 import java.util.List;
 import javax.ejb.Remote;
-import tn.esprit.overpowered.byusforus.entities.users.Candidate;
 import tn.esprit.overpowered.byusforus.entities.users.CompanyProfile;
+import tn.esprit.overpowered.byusforus.entities.users.Professional;
 
 /**
  *
@@ -29,7 +29,7 @@ public interface CompanyProfileFacadeRemote {
 
     List<CompanyProfile> findRange(int[] range);
     
-    List<Candidate> getSubscribersList(Long idAdmin);
+    List<Professional> getSubscribersList(Long idAdmin);
 
     int count();
